heterodyne: Make better provision for sloping baselines.
In some recent (20140520 #30-32) data there was an appreciable slope
in the baselines.  Without this extra step the initial historgram
bounds were too extended and it lead to large sections of baseline
being omitted (masked) from the non-linearity tests.  After this
change the initial histogram delineates the peak from the emission
more effectively, and there is better convergence to a sensible
peak value.  The result is the emission line is masked not portions
of the baseline.