heterodyne: Increase maximum non-linearity threshold.
It's another of these arbitrary where to set the bar fudge constants.
Rather than an absolute limit I've applied a maximum above the mode.
This may work well for some recent data the exhibits a new kind of
hybrid sloping sections along the spectral axis comprising
high-frequency noise.  The effect is to raise the global non-linearity
measure such that the previous 0.01 threshold was too low. As to
whether this threshold is now too high for other data I do not know,
and requires more testing.  It ought to be reasonable as it's a
constant above the mode.

An example of this noise can be viewed in a20140520_00032_01_0001, H08.

_FLAG_HIGHFREQ_INTERFERENCE_SPECTRA_ doesn't deal with this new strain
of h.f. noise yet.  If that were changed it may have an impact here.