heterodyne: Prevent histograms with negative sums.
FINDBACK normally tracks below the curve, but it can track above.  When
the histogram 'background' is subtracted, when arguably there is none,
it leads to negative bins and consequently may form a total count that
is also negative.  Taking the square root of this for the clump finding
rms causes the pipeline to exit.  This change uses MATHS to ensure
there are no bins with negative counts.  In practice this bias will not
matter because clump finding is likely to fail and the peak bin is
used.