PICARD: mask out low variances here, update POD
Previously the low-variance masking was a step in the mosaicking
recipe. However, since it is only a stopgap to deal with maps created
before the new "hitslimit" config parameter was added, and as such
should only ever be optional, it is probably better to control it from
within the current primitive, rather than have it as a standalone step
and only executed if a recipe parameter is set.