SCUBA2: alter daisy filtering for bright_extended
Changed the filtering parameter which the pipeline adds for daisies
reduced with bright_extended from filt_edge_largescale=200
to flt.filt_edge_largescale=200, based on recent email correspondence
indicating that the wrong parameter appeared to have been used.

Also added a warning so that it can be seen when this is done.