general: new primitive to check history component for given task
Read the history component using the NDF library and step through any
existing history records until a match is reached fr the given task
name. The task name may be a Starlink task or an ORAC-DR primitive
name. The match is a comparison, not an equality check, so partial
names are permitted. This may or may not be problematic. If in doubt,
be specific.

Consider this primitive as experimental as it has not been
exhaustively tested. It handles the presence of a valid history and
checks for the lack of history. Other errors are untested.