SCUBA2: store fastflat info for QA
The initial setup always appears to fail the Nbolo QA check, but this
is only due to the results not being stored correctly. In one case the
data weren't being stored at all; in another they were being
overwritten as the hash containing the results was re-initialized from
scratch each time.