SCUBA2: return calibrated files to the current object
In the pipeline, the calibrated data files are uncropped, and it is
these files that should be stored as reduced products, so they can be
used for science, rather than the cropped files.

In PICARD, the user is supplying their own science data which may
already be calibrated - any files calibrated by the recipe are usually
deleted.