heterodyne: Convert PARAM2 to pixels for UNMAKECUBE.
This involve adding some PIXSIZE arguments to pass the pixel scale to
UNMAKECUBE.  Only use the original WCS scale conversion if PIXSIZE is
undefined.

This hadn't bitten us before because we normally choose Gauss or
Nearest as the resampling method.  Without the correct scaling, the
size passed from the recipe parameter in arcseconds, convert to pixels
could be many times too large, and the UNMAKECUBE command would
timeout (or run for days).  In fact PARAMS2=4 seems as large as you
can go with ORAC-DR timing out on the JAC high-performance desktop
computers.