general: set ORACUT in the hdr if not set, using DATE-OBS
In some cases the ORACUT entry in the Frame or Group hdr may be
undefined, because the UTDATE FITS header is missing from the
file. This happens, for example, when SCUBA-2 images from different UT
dates are coadded, which causes calibration attempts to fail. Without
further information, the simplest option is to set a dummy ORACUT
using the UT date of the DATE-OBS which should always be defined.