SCUBA2: trap undefined FCF, set ORACUT and try again
Calibration of coadds/mosaics made from data from different UT dates
will fail because the ORACUT header is not defined. Furthermore, the
FCF returned by the calibration object in this case is 0 leading to
empty (zero-filled) maps.

Now issue an error that ORACUT was not defined and inform user that
calibrating coadds is not recommended. However, proceed with setting a
suitable ORACUT from the DATE-OBS and try again to retrieve an
FCF. This time it should succeed and the calibration can proceed.

Most of the change in this commit just comes from wrapping the
calibration step in a check for a defined FCF, which issues an error
if none could be determined.