scuba2: custom config for M14AH16C to deal with blobs
The default bright_extended reduction for this project yielded intense
blooms (or blobs) in the final maps, not always around the edge. This
config adds the "fix_blobs" extension to try and deal with them. It's
not perfect but it definitely helps. Hopefully the parameters can be
refined by the PI.