WFCAM: New recipe for fast-moving asteroids
This is variant of the JITTER recipe, whereby the relative base
positions are used to apply offsets to the WCS, to enable the mosaic
to show the asteroid as a point source, rather than trailed.  A
consequence of this is that the accurate astrometry cannot be derived
and the mosaic has an additional SKYBASE WCS Frame.