SCUBA2: add NOCHUNKING
This parameter determines whether or not the pipeline aborts after
checking if the data need to be split into chunks for
processing. Default value is 0 (do not check). Currently if any
non-zero value, the pipeline will abort only if the data need to be
chunked (i.e. memcheck = 2 is not yet supported).