heterodyne: Add headers to structure the commentary.
The ACSIS ORAC-DR logs are quite verbose and it's not always clear
what's going on in the big picture.  this commits adds some headings
to attempt to break up the logs into clearer sections.  I have
adjusted the numbers of underscores for underlining these headers
based upon the X display.  They will appear too long in the
.oracdr_<pid>.log file, but if I used the length of the heading
string, the X display underlines would look too short.  The latter
looks worse than the former.

There are some more to do, but this addresses the main primitives.