ifu: Improve mosaic-making efficiency.
Rather than create a mosaic from scratch for all of the cubes made thus
far, form a mosaic from the current mosaic cube with the pair of cubes
from the current quad.  The variance should take care of the weighting
although the zero pint would be better processing all the data.  Under
the former arrangement for n quads there would be 2*n*(n-1) operations
for MAKEMOS; under the new arrangement it's 3n.

Also the WCSFRAME calls that bracket the MAKEMOS step are invoked one
for each quad rather on average n.

To do this an extra steering header QUAD_CYCLE is added to select the
subgrp for the current quad.