heterodyne: Report velocity range for the 90% summation.
In the no-retrieve mode or when the limits were not recovered, rather
than write out zeroes for the velocity limits in the log.standard,
instead use the actual limits.  In practice they will be
distinguishable from limits obtained from the standards' table because
they will not be round numbers and have two decimal places.

These should be distinct from the velocity ranges read from the