heterodyne: Combine flats on a given UT date.
While this may not always be perfect, more often than not it should
improve the quality both in signal-to-noise ratio, and systematic
effects due to lack of uniform flux contributing to each receptor.

It would be good to save in a calibration index accurate ratios
derived from bright and extended emission, and reuse those for
observations where the flux is too weak or covers too small an area
for a self flat.