SCUBA2: fall back on date-obs/end FITS headers for elapsed time
PICARD doesn't know about the DURATION_SCIENCE uhdr, so just use the
DATE-OBS and DATE-END FITS headers to get the elapsed time.

In addition, a different file has to be chosen when called from
PICARD, assumed to be as part of the CHECK_RMS recipe. PICARD does not
know about "reduced" images, so use either the "calib" or "newimage"
data depending on whether or not the data needed calibrating before
analyzing.

The usual caveat applies that this will given nonsense results for
coadds/mosaics.