SCUBA2: only check for two-component fit if requested
Testing has shown that two components fits are generally poorly
constrained and do not yield significantly better estimates of the
beam size and/or FCFs. Therefore, they are henceforth disabled by
default and only available on explicit setting of the FIT_GAUSSIAN
recipe parameter to 2. This limits the use of two component fits to
PICARD recipes as the main pipeline does not support this parameter.