SCUBA2: new primitive to estimate error beam fraction
This primitive uses aperture photometry to measure the fraction of the
total flux that lies outside a FWHM-sized aperture. The value is
printed to the screen by default.

Ideally this would be achieved with a single call to autophotom, but
that would require a specialist modification to the
_APERTURE_PHOTOMETRY_ primitive. Instead, use two calls and ensure
that the initial position and sky/background annulus are exactly the
same for each call.

The result is stored in the calibration system using the errfrac method.