SCUBA2: add CALCNEFD argument to avoid unnecessarily recalculating NEFD
This new input argument determines whether or not the NEFD image
should be recalculated for consistency after the matched filter has
been applied to an image. The default is true and it should be true
for cases where the matched filter is an output product, but it should
be false for those cases when the matched filter is only used to help
with peak finding.