heterodyne: change _FIND_TRUE_PEAK_ to work for stares/grids/jiggles
For grids and jiggles, this will now just use (1,1) for the peak if it
contains valid data, otherwise it will use the peak pixel of the
integrated map. This primitive should be merged with code that does
the same thing in _CREATE_MOMENTS_MAPS_THROUGH_SMOOTHING_, but I need
to check if the difference in recipe parameters will cause anyone
problems.