SCUBA2: keep track of integration time in the summit pipeline
The duration of the data-taking part of the observation is stored in
the Frame uhdr. An easy way to keep track of the total integration
time is to just keep adding the numbers. This will result in a
per-object total integration time if maps are coadded across
observations.

Occasionally this may overestimate the total time in those cases when
the pipeline picks up data from some but not all of the subarrays.