heterodyne: Check and report each hybrid subband independently.
For hybrid data it is possible to have one subband mostly bad but
others not.  If combined into one large dataset, as if concatenating
subscans, the badness is diluted, and so the receptor's bad subscan is
retained.  This can cause problems with merging, as such data seem to
have weird non-bad perpheries (cf. 20140210 Obs 20 Receptor H01
(Subband 1) & H06 (Subband 2), and 20140213 H01 (Subband 1).

For now at least take the approach of excluding the entire receptor if
one of its subbands is mostly bad.  Adjacent good receptors could be
retained provided the suspect subband is made entirely bad before
merging is attempted.