heterodyne: Add FRACTION argument.
This is also a recipe parameter FRACTION_BAD.

Adding this tuning proved necessary because some suspect hybrid
subscans have around 80% bad values with other crazy patterns and
values outside of the bad values.  The artefacts include alternating
bad stripes at the ends (but extending inside the normal noisy edges
of spectra), and a period of strong non-linearity.  These need to be
tackled before merging otherwise artefacts appear in the composite
spectra.  So the normal 0.999 it far too tight.  I had thought that I
had reduced this to 0.9 or at 0.99 months ago, but now do make
the default maximum fraction a sensible 0.9.