calib: Allow combinations of a list and another option for bad_receptors.
Jane Buckle wanted to supplement the default bad-receptor method on
the command line, not replace it, when specifying a list of receptors
to exclude.  The same colon-separated syntax is used.  For instance it
is now possible to write "-calib bad_receptors=H01:H06:index" to
exclude H01, H06, and any others found in index.bad_receptors_qa.