SCUBA2: log.mapstats has columns for rms and nefd units.
log.mapstats files create by picard will correctly write the units
into the header of the table. However, if it is run on a second
file with different units it doesn't indicate the change in units.
This commit moves the RMS and NEFD units into columns in the table.
It also adds in a column for the RECIPE fitsheader.