SCUBA2: print filename in log.mapstats entries.
As new log.mapstats calls have their results concatenated into an
existing file, it can be easy to not notice if you've run it on the
wrong file. Adding in the filename to the results should help a little
for users who are running this manually with the PICARD recipe rather
than in the ORAC-DR reduction pipeline.