heterodyne:  Form two hashes for the relative signals.
One is for the full set of defined receptors, and a subset
without any bad receptors which is used for the pretty printing.
The loop is now for all receptors, not just the good ones,
whose number may vary between observations.  The full set will be
used to apply the normalisations.