heterodyne: Form normalisation array for each observation.
When I re-jigged the code to work for more than one observation at a
time (to improve the S/N), I created arrays with the maximum number of
receptors (16) in case different observations had different sets of
available receptors.  Now when dividing by the relative responses at
the end the primitive used the full 16 receptors instead of the actual
number present in the original data.  This could lead to the wrong
factors being applied to each receptor.

The cure for this applied here is to select just those receptors
applicable to each observation and use their flat-field factors to form
a temporary NDF with TRANDAT.  Then this is grown to a cube as before,
and divided into the corresponding observation file to apply the flat.