JSA: allow use of JSA tile naming in catalog recipe
Most of the JSA_CATALOGUE recipe has been moved to a primitive
_CREATE_JSA_CATALOGS_ in order to allow it to set up suitable
frame and group classes.  This is required in order to allow us
to use JSA methods such as inout_jsatile and jsa_pub_asn_id.

Also changed from running the primitives "in frame mode" to running
them "in group mode" because we need the "group mode" kind of headers
in order to be able put the catalogs in the "public" category
(instead of "obs") of the JSA.

The change to the classes prevents FITS files from being added
to the object, so the FITS catalog file names have to be
recorded in the uhdr instead.