JSA: set product ID prefix for extent and peak
The extent plane will contain multiple products, so construction
of the product ID by appending the subsystem to the product is
unsuitable in this case.  Since access to the product-overriding
argument of the product_id method is not always available, (e.g.
when that method is called by collate_headers), it appears that it
is necessary to be able to store a product ID prefix separately from
the product name in the object.  The extent and peak catalog
creation primitives now set the product ID prefix in the object
and unset it again before they complete.  Additionally collate_headers
must no longer attempt to force the product part of the product ID
to be the value returned by the product method because the product_id
method already falls back to the product method so it isn't necessary,
and when it does do so it prevents the product_id method from making
use of the new product ID prefix setting.