JSA: move peak cat preview preparation code
Moved the code which prepares to create preview images for the
peak catalog to the _CREATE_JSA_PEAKS_CAT_ primitive.  This is because
within this primitive we have the correct product ID prefix context
and because primitives normally handle their own preview images
(via one of the JSA tagging primitives).  This is in preparation
for adding the correct product ID to the peak catalog previews.