SCUBA2: fudge pointing positions in JSA tiles
The _FUDGE_CALIBRATOR_POSITION_ primitive now has a POINTING argument
which causes it to also attempt to apply pointing offsets for pointing
observations.  This option has been enabled in the
REDUCE_SCAN_JSA_PUBLIC recipe.  Note that _MAKEMAP_FRAME_ also checked
that the source is a recognised flux calibrator before calling
_SHIFT_IMAGES_, so it was necessary to add an argument, SHIFT_NON_CAL,
to disable this check.