SCUBA2: use SkyRef when fudging pointing position
The _SHIFT_IMAGES_ primitive only used the SkyRef attribute if
the source was a recognised calibrator, which meant that
_FUDGE_CALIBRATOR_POSITION_ would not get offsets for non-calibrator
pointing observations.  Therefore a new argument USE_SKYREF
has been added to _SHIFT_IMAGES_ to instruct it always to use
the SkyRef position.  This argument is passed from
_FUDGE_CALIBRATOR_POSITION_ through _MAKEMAP_FRAME_ via a
new argument SHIFT_USE_SKYREF of that primitive.