SCUBA2: make failure to fudge pointings fatal
_FUDGE_CALIBRATOR_POSITION_ now calls orac_termerr if it is
processing a pointing observation and is not able to determine
pointing offsets.  This is because pointing observations would
tend to be carried out when the pointing needs to be corrected,
and we do not want to include poorly pointed pointings in JSA
HEALPix tiles because they may smear the image.  Therefore
the recipe should fail when trying to reduce a pointing and the
position cannot be fudged.  The alternative would be to mark
these observations somehow for subsequent exclusion from
co-adds.

The addition of this check was agreed by JTD.