heterodyne: Have arguments to convey user-requested baseline regions.
This is applicable to spectra with weak emission and multiple
lines.  Some users prefer to specify where they expect no emission
rather than rely on automatic mode in MFITTREND to mask all the lines.

It adds a new BASELINE_REGIONS argument to ITERATIVE_GROUP_PRODUCTION_,
from which lower and upper bounds are passed to _REMOVE_BASELINE_
through new LOWER and UPPER arguments.  There is a corresponding
BASELINE_REGIONS recipe parameter.

Also included is a reordering to alphabetical of the METHOD/ORDER
arguments in _REMOVE_BASELINE_.  [This change could not be easily
disentangled from the new LOWER and UPPER arguments.]