heterodyne: Extend BASELINE_REGIONS to non-linearity tests.
If the user wants to define the locations of baseline, it seems silly
not to use them also for baseline non-linearity tests.  One could
detect BASELINE_REGIONS being set and override BASELINE_LINEWIDTH, but
for now I have added a further option for the latter recipe parameter,
namely "base" to choose the BASELINE_REGIONS.  As the two non-linearity
primitives mask the emission first, the complement of the
BASELINE_REGIONS must be formed before performing the QA tests.