heterodyne: Expand the noisestats table
Add additional columns for the UT start time, observation number,
exposure time, airmass, tau, and the RMS units and label.

Insert a hash comment symbol at the start of the column names.
Also add a comment giving the creation time of the log.

At present the format is such that column names align with their values,
but we may need to switch to CSV in the future for better machine
analysis.