SCUBA2_MAPSTATS: use scientific notation for rms and nefd
The correct number of decimal places to print to is very dependent
on the units. Easiest solution is to use scientific notation.

Also now reads the nefd units from the nefd map, as if users
manually calibrate their data with cmult/setunits they will not
change the units of the nefd extension.