SCUBA2: make jsa_generic the default dimmconfig for REDUCE_SCAN
This commit attempts to ensure that all other current recipes keep
their existing behaviour, as do the summit and quick look pipelines,
but that the default for future recipes that don't specify anything
will be to use jsa_generic.

In general, the type of makemap-config used is set in _SCUBA2_HELLO_
for a few specific observation types, and is left unset for all other
types.  If a CONFIG_TYPE parameter is given when calling
_MAKEMAP_FRAME_ this will override anything set in _SCUBA2_HELLO_ (so
this is used in recipes like REDUCE_SCAN_FAINT_POINT_SOURCES that want
to use a specific dimmconfig for all observations). This can always be
overridden if the user requests a change through recpars.

The intent of the changes being made:

SCUBA2.pm: change name from "normal" to "base" to help make more
explicit the difference between our default, generic dimmconfig
'jsa_generic' and recipes that specifically want the "base" config
'dimmconfig.lis'. Also updated _SCUBA2_PCA_ to use "base" instead of
"normal"

_SCUBA2_HELLO_ and _MAKEMAP_FRAME_ have been altered so that they now
do not set the Frm UHDR 'MAKEMAP_CONFIG_TYPE' if its not already set.

_GET_MAKEMAP_CONFIG_: this now uses the "base" config as the default for
the QuickLook and Summit pipelines. For other pipelines, it uses
'jsa_generic' as the default.

The recipes 'REDUCE_SCAN_CHECKRMS', 'REDUCE_SCAN_FAKEMAP',
'REDUCE_POL_STATE' and 'REDUCE_SONS' have been changed to use the
'base' dimmconfig as a default, replicating their current
behaviour. Someone who knows what these recipes are used for should
decide if this is appropriate, or if they should use jsa_generic (or
be removed). It is believed that all other recipes apart from
REDUCE_SCAN either do not use the MAKEMAP_CONFIG or specifically
specify a config_type themselves.