heterodyne: Stop the threshold from growing too much.
When it grows it is usually a sign that there is some crap data
biasing the threshold, and hence sensible data being rejected, further
raising the bar in a runaway.  The 10% increase is a bit arbitrary but
you would not expect after an iteration correcting for the first-order
receptor sensitivities that the mean threshold would rise by more than
this.  In the light of experience it might need adjustment to 12 to 15%.