heterodyne: Count your blessings in point-source catalogue.
For reasons I don't unsderstand, the $result->misc method barfs that
$result is not blessed, but Data::Dumper of it suggests that it is.
It begins like this.

$VAR1 = [
          bless( {
                   'Y' => undef,
                   'MORPHOLOGY' => undef,
                   'SPECTYPE' => undef,
                   'FLUXES' => undef,
                   'MISC' => {
                               'flux850' => '31.2'
                             },

This code I copied from _QA_SENSITIVITY_VARIATION_ (and it's
elsewhere) so you would have thought that it has worked.  Given that
there are not many entries with a line width, needed here to mask the
line.  I've decided to remove this refinement, but would like to come
back to it in the future to understand what' wrong when there's less
time pressure.