heterodyne: Factor out code to paste subscans for a single receptor.
This mostly originated in _INDEX_BAD_RECEPTORS_.  This has extra code
offering an efficiency gain of a single PASTE command without NDFCOPY
commands when all the subscans are of equal length, i.e. have the same
number of spectra.

Also looking to use it elsewhere to the high-frequency noise filtering
faster for multi-subscan observations.