heterodyne: Fixed bugs for multi-subband observations.
Not sure what I was thinking when I originally put this primitive
quickly together.  For subbands it was reporting whether the data were
good after each subscan, and while pasting it was using the wrong
offset (the end rather than where the last subscan ended) thus giving
wide bands of bad values between, and as a consequence giving rise to
failures.  So even if the first subscan reports as good, for each
subsequent one the originl primitive would report a failure, thus
rejecting most or all receptors.  The second bug is corrected in the
new _PASTE_SUBSCANS_BY_RECEPTOR_ primitive all the data for a given
receptor.  This primitive also has efficiency benefits too.