heterodyne: Fix bug when a receptor is rejected.
By not using the correct object to test there could be a blank line in
the file for TRANDAT when a receptor had been deemed bad during QA.
TRANDAT interpreted this as the end of file and died, since this was
encountered before the total number of values it expected.  This
caused JAC Recipe Instance 66575, for example, to fail.