heterodyne: Restructure high-frequency interference filtering.
Rather than process each subscan individually and each receptor within
in that subscan, this restructuring (akin to what's already in the
low-frequency noise filtering) combines all the spectra for a receptor
and analyses it as a whole, including ringing.  This has two main
benefits.

-  It's somewhat faster. The gain depends on the number of subscans,
but for an example with seven subscans, doing it this way was three
times faster.

- The analysis of the edginess profile is more robust.  The
determination of the base level is better defined, and cases where
extended ringing persists between subscans are better defined.

There is also a loop for subsystems.  This has not been checked since
merging now takes place before the filtering, but the code is
essentially as used in _FLAG_LOWFREQ_INTERFERENCE_SPECTRA, which was
tested before the change to merge earlier.