heterodyne: Protect FIXSTEPS if there is more than one subscan.
This is needed because the pasting performed by
_PASTE_SUBSCANS_BY_RECEPTOR_ does not manipulate the JCMTSTATE
extension, which should be concatenated to the number of spectra in
the pasted NDF, but also sliced for the single receptor.  FIXSTEPS
complains with an HDS error "DAT__SUBIN: Subscripts invalid".

Another solution is to use TIMESORT in _PASTE_SUBSCANS_BY_RECEPTOR_,
which does perform this manipulation, but at the moment this is not
delivering non-bad data arrays except for the first receptor.

This change might not give the best results in that FIXSTEPS filter
will be absent for many reductions, but it will prevent the many
confusing HDS errors that might concern a user; and which also
confuses the JAC processing system, which reports DAT__SUBIN as the
error rather than the later processing step that has actually caused a
job to go to the Error state.

If TIMESORT is made to work then this commit can be undone.