heterodyne: Add JCMTSTATE argument to the subscan pasting.
This is to allow TIMESORT to do the pasting and manipulation of the
JCMTSTATE extension's arrays, as needed by FIXSTEPS.  At present the
TIMESORT step fails to deliver non-bad data arrays in a consistent
fashion.  Part of the change needed was to match the spectral bounds,
as TIMESORT only warns about this and generates a bad data array, but
these warning sare only seen in ORAC-DR if -verbose is enabled, and so
fails silently.