JSA: generate "extent" preview for non-detections
Following the JSA meeting 1/13/15 decision to keep the current
plane arrangement for JSA catalogs (see Trac ticket 482) it is
necessary for there to be a preview for "extent" planes in the
non-detection case.  This commit aims to generate such previews
by overlaying the text "No emission detected." on the normal
"extent" plane greyscale background.

This required adjusting the _TAG_AS_REPRESENTATIVE_ PREVIEW_BACKGROUND
parameter for the case when there is only the background and adding
support for rendering text to _CREATE_GRAPHIC_FROM_FILE_.