Docs: automatically generate recipe info for 260,264&265
Now requires pod2sst to be in the path. Starlink now includes
this, but users will need to build the perlmods.

This uses pod2sst and prolat (in starlink) to generate SST style
documentation from the PODs in each recipe. It also attempts to
categorize the recipes in a useful way. SUN/264 (SCUBA-2) uses a
'nonscience_recipes' file to separate out the recipes for non-science
observations.  Something similar should be added into the heterodyne
(SUN/260) document.