Update to SUN269 version 0.4:
sun269.tex: Expand Introduction. Add more running and setup details, along with information on how logging works and where the files are. Add more information to Appendices C and D to specify where recipes and primitives are searched for and in what order and how to control or modify this.
sun269/primitives.tex: Add some missing primitives.