spectroscopy: Handle change to EXTRACTOR's FITS header searching.
UKIRT FITS airlocks are not terminated by an END card, but
EXTRACTOR uses the END header to end its search for a named keyword.
In this case the header is SATURATE.  For the latest KAPPA, insert an
END header before running EXTRACT, otherwise write the SATURATE header
with the SATUR_LEVEL value from the Sextractor config file.