spectroscopy: Kludge to get around mysterious non-closing graphics.
The _ss histogram display in the GWM window appears not to be closed.
Trying in ICL it was not reproduceable, and indeed there is a close
in HISTOGRAM and it doesn't effect all histogram plots.  This seems
the least intrusive of a number of working options.

Should return to this when there's more time in case it comes back to
bite us.