SCUBA2: explicitly crop before calculating MAPSTATS
This allows SCUBA2_MAPSTATS to work on maps reduced using an external
mask, or otherwise where 0,0 isn't the center of the map. It will
extract a rectangular region from the map center, defined by the FITS
headers MAP_HGHT, MAP_WDTH and MAP_PA.

This is done in _WRITE_MAPSTATS_LOGFILE_ instead of in
_FIND_IMAGE_STATISTICS_, as it is easy to crop the whole file and all
of the smurf extensions (e.g. nefd, exp_time) at once, but hard to
pass only the extension data array in and crop it sensibly.