heterodyne: Add fudged protection against weak extreme peaks.
The first (or last peak for negative skewness) may not be the one
required.  It is possible to have a weak peak in a tail of the
histogram in the opposite sense to what the skewness indicates.  It's
not common, but will lead to the wrong thresholds being set and
later to an all-bad spectrum from baseline fitting causing the
pipeline to exit.  This is what occurred in JSA processing job 140874.

This restructures the CUPID catalogue parsing to find the strongest
peak first.  If EXTREME=1 and the first (or last peak if negatively
skewed) has less than a tenth of the signal of the strongest peak,
the strongest peak is selected instead.

Also fix bug to start searching for the strongest peak from the line
_after_ the BEGINTABLE.