heterodyne: Fixed bounds bug in flat-fielding.
The dimensions of the last file being processed was being used to grow
the ratios into a cube.  Now in most cases these will be the same,
when something goes wrong and an observation is interrupted then
resumed in a new observation, the time bounds of the time-series cube
will change.  So remove the factorisation introduced for speed and
obtain the shape of each input cube before growing the ratios.