heterodyne: Choose most populous peak if it's over 3x more than first/last peak.
When I introduced a fudge to cope with histograms that had an unusual
combination of skewness and and extra peak (clump) on the opposite
side expected, I added a get out option to use the most populous peak
if its sum was more than 10x the first or last clump (depending on
skewness), but noted that the ratio was totally arbitrary and might
need to be lowered in the light of experience.  10x seemed a little
too high at the time.  For 20130705 Obs 9, which prompted this change
for JSA job 138140, a factor of 5 was nearer the mark.  Set to 3x and
hope that it's not gone too far the other way.  Re-processing all the
legacy data should help determine the correct factor.