heterodyne: Modify logic structure for narrow overlaps.
Not sure how the code worked when the overlap for the non-noisy ends
is smaller than the minimum.  Adjust to have an elsif clause rather
than a separate if block.  This was needed not only for the case where
a line occurred in the overlap, but now also noted is the case of
early observations where the overlap was too narrow for the noise at
the overlapping ends of each subband.  This allows the recipe to
complete, but the results may be dreadful (too few comparison points
affected by noise).  Some other technique like diferencing the
baseline levels in adjacent subbands should yield better results.