general: Cope with the sparse array case for images.
Small thumbnail images of HEALPix of small spatial regions are sparse
and may have too few valid data points to estimate standard deviations
for the default Faint scaling in DISPLAY.  For these (like zero-range
data) use the Scale mode so that DISPLAY does not exit with an error.
There's is so littlei information in such thumbnails, the choice of
scaling is not going to matter.