imaging: Shrink search window to avoid selecting the other beam of the same polarity.
The centroid search could start at the other beam of the same polarity
given the right combination of chop and throw sizes.  Trim off a few pixels
to avoid this yet retain a large search window.  The triming might
still be too small in some cases.

The effect was to make the _cab images too thin along one axis, and thereby messing
up the photometry.