PARSER: improve debug output
When the -debug option is given, oracdr will now print out the full
list of arguments into the log. Although this makes the debug output a
little uglier, I think the usefulness of this is more important. The
ORACDR.DEBUG file  frequently has many calls to the same command,
and matching them to the the calls in the log is finicky.