SCUBA2: raise no errors if attempting to register faint sources
Fix the bug that caused an error if the source was faint. Instead, if
the register images fails because the source is too faint, simply warn
the user what has happened and continue on.

Also update the messages to user in _MAKE_MOSAIC_GROUP_BATCH_ so
that it is clear to the user what is going on.