heterodyne: remove tileborder from _CREATE_CUBE_FRAME_
The history of this recpar is very confused. This was previously
removed from _CREATE_CUBE_GROUP_ but none of the calls to it were
updated to stop setting the tile border. The TILEBORDER param appears
to have been initially added to allow smoothing at the edges to work
well, but at a later point it was removed from the group processing
and left in for the obs level processing. This commit will make the
behaviour in the FRAME level products the same as in the GROUP
products, i.e. no tileborder is included. Presumably this no longer a
problem, or this parameter wouldn't have been removed from
_CREATE_CUBE_GROUP_, as far as I can tell from the code.

This was done to solve a JSA problem, as the JSA processing code
now expects there *not* to be a bb mask in the final files.