general: Replace clashing equals sign in the astro:name keyword=value.
Some OBJECT names include an equals sign, e.g. for co-ordinates of a
tile in a survey.  The extra = causes a problem later in JSA
processing (JSA::Convert) when the EXIF keyword=value headers are to
be split into their component parts.  The solution was not obvious.
For now I've replaced the symbol with the word equals.  Another option
would be to substitute a colon.