heterodyne: Derive and store the extreme values of the representative spectra.
These are useful data for constaining limits in other primitives, such
as for spike exclusion.  This is an interim solution as it only provides
the average over emision and background.  There could be much brighter data
emission.  Better in the longer term would be some filter looking for
anomalous values in the running median, albeit slower.