heterodyne: Improve spike filtering limits.
This is an interim solution to overcome marking bright emission as bad.
A new primitive is needed to flag noise spikes, with an option to set
the line width for better disrimination in recipes aimed at broad-emission
lines.  Data that fall foul of the median +/- 20 are rare.

It uses the internal headers MEAN_SPECTRA_MAXIMA and MEAN_SPECTRA_MINIMA
(derived from the mean spectrum of each subband) to set new limits
instead of the median +/- 20.  Also increase the range to extend from
these limits (or from the median if the headers are not defined) from 20
to 30.