heterodyne: Add support for chunking of HEALPix tiles.
A few percent of the maps in the HARP HPX legacy reductions fall foul
of memoery requirements and/or more than 2**31 pixels.  In order to
circumvent this, chunking is added for the JSA tiling.  Instead of 1
to ntiles the combination is performed for each unique HEALPix tile
identifier.  [Note sometimes MAKECUBE would claim to have generated
tiles (from its JSATILELIST output parameter), but not all were
present, so a list of actual tiles created were then reassembled from
the chunks.] In comparisons the integs look the same by eye, an
statistcally they are virtually the same in the mean.  However, there
are sometimes some linear(ish) features through the integ of a few
tenths.  The source of these is currently unknown.

The main changes were to _CHUNK_AND_MOSAIC_CUBE_.  This has new GROUP,
JSATILE, JSATILELIST, and SUFFIX arguments. an expanded description of
the output products.  It reports the HEALpix tiles created, which are
reasembled in increasing numerical order.

This had knock-on effects to _CREATE_CUBE_FRAME_ and
_CREATE_CUBE_GROUP_.  Since chunking had not been performed on
individual observations, only for the group cube _CREATE_CUBE_FRAME_
needed more-extensive changes to assess the chunksize and whether or
not chunking was required or demanded, and setting the names of
the output cubes.

_CREATE_CUBE_FRAME_ involved an API change (CHUNK and CHUNKSIZE
arguments) affecting _CREATE_CUBE_MASK_FRAME_FROM_GROUP_ and
_ITERATIVE_GROUP_PRODUCTION_.  The CHUNK boolean is to only chunk for
the legacy final products whose bounds may exceeed the 2**31 limit,
but never will in non-HEALPix tiling.

The thresholds to chunk are somewhat arbitrary as no clear method has
been identified that indicates when the number of pixels exceeds the
32-bit limit.  The number of HPX tiles is not the solution.  One
approach would be to call MAKECUBE with OUT=! and inquire the bounds
of the map, and hence derive the number of pixels, adding time-series
cubes in sequence until the limit is reached.

There was some fiddling to prevent single files being procesed one at
a time, since this was outside of the original design of the chunking
primitive, and would a problem on the last chunk.

This has been tested on small and larger jobs, normal and hybrid.
However, the easiest way to test it more thorough is in the legacy
pipeline where many more reductions can be checked.